.github/requirements.txt
jsonschema
pytest
pyaml
